# R9M using wrong Idle method and EEPROM speedup