Proposal for fixing the woes of R9M users, who have just the worst experience when changing things in Lua. TL;DR I believe the EEPROM commit on the TX takes a full second or more, during which time the RX just falls out of sync.

Based on my commit-telemskip branch (#1127).

Ref Issue #1130

### Problem 1: Using the wrong Idle method
The code is only supposed to use the `hwTimer.pause()` method on STM32 targets that use EEPROM emulation. However, the define check would use it even if there was external EEPROM. The #else code is only designed to work around interrupts stopping while flash write is happening. and therefore I2C EEPROM should still work fine but use the Idle timer callback. I'm pretty sure this was just a mistake in the definition that went unnoticed.

I'm fairly sure the code should be: "on every platform except STM32 without extEEPROM". That's what I've got now.

### Problem 2: EEPROM.write takes a long time
I don't have an R9M but I do have an R9MM with external EEPROM, and writing the 96 byte RX config takes 386ms. If I extrapolate that to the 268 byte TX config, that's an estimated 1078ms-- over 1 second of no packets going out! If I emulate that delay on ES900TX, I lose connection to the RX quite regularly when changing values via Lua on 200Hz. The RX just can't stay in sync properly for that many packets (which is sad, I wish it worked longer). It works better at 50Hz but still isn't perfect.

To reduce the delay I've changed two things:
* Increase the I2C speed from 100kHz to 400kHz. I'm not sure if this is out of spec for the EEPROM but it seems to work fine for my R9MM RX and YeOldePirate's R9M TX. The bus speed increase only reduced the write time on the RX from 386ms to 340ms. I've added the define `TARGET_EEPROM_400K` and applied this to only the R9M TX target.
* Only write the _changed_ bytes to EEPROM. The actual physical EEPROM write is what takes a long time, changing flash is hard, yo. For STM32 with external EEPROM, I've changed the code to use `EEPROM.update()` instead of `EEPROM.write()`. This does a read and only does a write if the value is changed. This potentially doubles the I2C traffic, but in practice only 1 or 2 bytes are changed at a time so that's roughly the same traffic except a ton faster because no actual writing is taking place. Reduces the write time from 386ms to 43ms, or 17ms at 400kHz.

### Testing
Tested with YeOldePirate, it seems they can't break the connection just by changing a Lua param any more. It still is possible to desync if there's a long delay with 0 packets, this does not address that which is a completely different issue.