@YeOldePirate 2021-11-24
yup this pr fixed my issues, i have been flying all day on it and it has worked flawlessly since flashing even on 868mhz